refactor: fix structure issues
