+++ app/page.tsx
@@ -1,7 +1,5 @@
 import type { Metadata } from "next";
-import Header from "../components/organisms/headerComponents/Header";
-import { PageWrapper, Main, StyledFooter } from "./page.style";
-
+import { Main } from "./page.style";
 // import Content from "./components/Content";
 
 export const metadata: Metadata = {
@@ -10,11 +8,5 @@ export const metadata: Metadata = {
 };
 
 export default function Home() {
-  return (
-    <PageWrapper>
-      <Header />
-      <Main>aaaa</Main>
-      <StyledFooter />
-    </PageWrapper>
-  );
+  return <Main>This is main page</Main>;
 }
