+++ components/hooks/useDevice.tsx
