+++ components/molecules/avatar/component/avatar.style.ts
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { sizes } from "@/styles/sizes";
 
-export const ProfileWrapper = styled.div`
+export const AvatarWrapper = styled.div`
   width: ${sizes.md};
   height: ${sizes.md};
   @media (min-width: 376px) {
