+++ components/organisms/curtainMenu/navbar/navbar.tsx
@@ -1,6 +1,6 @@
 import { menuData } from "./mobileNavbar/component/data";
 import { NavbarWrapper } from "./navbar.style";
-import NavList from "./mobileNavbar/component/components/navList";
+import NavList from "./mobileNavbar/component/navList";
 
 export default function Navbar() {
   return (
