+++ components/organisms/footer/components/footerData.ts
